[build-tools] place gradle profile step after run gradle
